nanoleaf2mqtt/pyproject.toml
[project]
name = "nanoleaf2mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomqtt>=2.4.0",
    "aiostream>=0.7.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "python-dotenv>=1.1.1",
    "stamina>=25.1.0",
    "zeroconf>=0.147.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
